requirements.txt
python-dotenv==1.1.0
fastapi==0.115.12
pandas==2.2.3
scikit-learn==1.6.1
openpyxl==3.1.5
xlrd==2.0.1
mlflow-skinny==2.21.3
boto3==1.37.34
uvicorn==0.34.1
gunicorn==23.0.0
# database operations
psycopg2-binary==2.9.10
sqlalchemy==2.0.40
alembic==1.15.2
# queue management
rq==2.3.2
redis==5.2.1
rq-scheduler==0.14.0
